+++ pkg/ebpf/tracer.go
@@ -6,7 +6,6 @@ import (
 	"bytes"
 	"expvar"
 	"fmt"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -82,13 +81,12 @@ const (
 	defaultClosedChannelSize = 500
 )
 
-// CurrentKernelVersion exposes calculated kernel version - exposed in LINUX_VERSION_CODE format
-// That is, for kernel "a.b.c", the version number will be (a<<16 + b<<8 + c)
-func CurrentKernelVersion() (uint32, error) {
-	return bpflib.CurrentKernelVersion()
-}
-
 func NewTracer(config *Config) (*Tracer, error) {
+	// make sure debugfs is mounted
+	if mounted, msg := util.IsDebugfsMounted(); !mounted {
+		return nil, fmt.Errorf("%s: %s", "system-probe unsupported", msg)
+	}
+
 	m, err := readBPFModule(config.BPFDebug)
 	if err != nil {
 		return nil, fmt.Errorf("could not read bpf module: %s", err)
@@ -161,11 +159,6 @@ func NewTracer(config *Config) (*Tracer, error) {
 		} else {
 			fmt.Errorf("error enabling DNS traffic inspection: %s", err)
 		}
-
-		if !util.IsRootNS(config.ProcRoot) {
-			log.Warn("system-probe is not running on the root network namespace, which is usually caused by running the " +
-				"system-probe in a container without using the host network. in this mode, you may see partial DNS resolution.")
-		}
 	}
 
 	portMapping := NewPortMapping(config.ProcRoot, config)
@@ -205,30 +198,6 @@ func NewTracer(config *Config) (*Tracer, error) {
 	go tr.expvarStats()
 
 	return tr, nil
-}
-
-// snakeToCapInitialCamel converts a snake case to Camel case with capital initial
-func snakeToCapInitialCamel(s string) string {
-	n := ""
-	capNext := true
-	for _, v := range s {
-		if v >= 'A' && v <= 'Z' {
-			n += string(v)
-		}
-		if v >= 'a' && v <= 'z' {
-			if capNext {
-				n += strings.ToUpper(string(v))
-			} else {
-				n += string(v)
-			}
-		}
-		if v == '_' {
-			capNext = true
-		} else {
-			capNext = false
-		}
-	}
-	return n
 }
 
 func (t *Tracer) expvarStats() {
@@ -283,7 +252,13 @@ func (t *Tracer) initPerfPolling() (*bpflib.PerfMap, error) {
 				if t.shouldSkipConnection(&cs) {
 					atomic.AddInt64(&t.skippedConns, 1)
 				} else {
-					cs.IPTranslation = t.conntracker.GetTranslationForConn(cs.Source, cs.SPort, process.ConnectionType(cs.Type))
+					cs.IPTranslation = t.conntracker.GetTranslationForConn(
+						cs.Source,
+						cs.SPort,
+						cs.Dest,
+						cs.DPort,
+						process.ConnectionType(cs.Type),
+					)
 					t.state.StoreClosedConnection(cs)
 				}
 			case lostCount, ok := <-lostChannel:
@@ -399,7 +374,13 @@ func (t *Tracer) getConnections(active []ConnectionStats) ([]ConnectionStats, ui
 				atomic.AddInt64(&t.skippedConns, 1)
 			} else {
 				// lookup conntrack in for active
-				conn.IPTranslation = t.conntracker.GetTranslationForConn(conn.Source, conn.SPort, process.ConnectionType(conn.Type))
+				conn.IPTranslation = t.conntracker.GetTranslationForConn(
+					conn.Source,
+					conn.SPort,
+					conn.Dest,
+					conn.DPort,
+					process.ConnectionType(conn.Type),
+				)
 				active = append(active, conn)
 			}
 		}
@@ -438,8 +419,13 @@ func (t *Tracer) removeEntries(mp, tcpMp *bpflib.Map, entries []*ConnTuple) {
 	for i := range entries {
 		err := t.m.DeleteElement(mp, unsafe.Pointer(entries[i]))
 		if err != nil {
-			// It's possible some other process deleted this entry already (e.g. tcp_close)
+			// If this entry no longer exists in the eBPF map it means `tcp_close` has executed
+			// during this function call. In that case state.StoreClosedConnection() was already called for this connection
+			// and we can't delete the corresponding client state or we'll likely over-report the metric values.
+			// By skipping to the next iteration and not calling state.RemoveConnections() we'll let
+			// this connection expire "naturally" when either next connection check runs or the client itself expires.
 			_ = log.Warnf("failed to remove entry from connections map: %s", err)
+			continue
 		}
 
 		// Append the connection key to the keys to remove from the userspace state
@@ -681,4 +667,10 @@ func SectionsFromConfig(c *Config, enableSocketFilter bool) map[string]bpflib.Se
 			Disabled: !enableSocketFilter,
 		},
 	}
+}
+
+// CurrentKernelVersion exposes calculated kernel version - exposed in LINUX_VERSION_CODE format
+// That is, for kernel "a.b.c", the version number will be (a<<16 + b<<8 + c)
+func CurrentKernelVersion() (uint32, error) {
+	return bpflib.CurrentKernelVersion()
 }