+++ pkg/util/containers/image.go
@@ -10,6 +10,13 @@ import (
 	"strings"
 )
 
+var (
+	// ErrEmptyImage is returned when image name argument is empty
+	ErrEmptyImage = errors.New("empty image name")
+	// ErrImageIsSha256 is returned when image name argument is a sha256
+	ErrImageIsSha256 = errors.New("invalid image name (is a sha256)")
+)
+
 // SplitImageName splits a valid image name (from ResolveImageName) and returns:
 //    - the "long image name" with registry and prefix, without tag
 //    - the "short image name", without registry, prefix nor tag
@@ -18,7 +25,10 @@ import (
 func SplitImageName(image string) (string, string, string, error) {
 	// See TestSplitImageName for supported formats (number 6 will surprise you!)
 	if image == "" {
-		return "", "", "", errors.New("empty image name")
+		return "", "", "", ErrEmptyImage
+	}
+	if strings.HasPrefix(image, "sha256:") {
+		return "", "", "", ErrImageIsSha256
 	}
 	long := image
 	if pos := strings.LastIndex(long, "@sha"); pos > 0 {